requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
pymysql==1.0.3
cryptography==40.0.2
pandas==2.0.0
matplotlib==3.7.1
openpyxl==3.1.2
Werkzeug==2.2.3
python-dotenv==1.0.0
gunicorn==20.1.0
chardet==5.2.0
# 仅在导出MySQL数据时需要
PyMySQL==1.0.3
